improved tests

+++ src/task4/calculator/Sign.java
@@ -1,7 +1,7 @@
 package task4.calculator;
 
 public class Sign {
-    protected int getSign(int n) {
-        return 0 == n % 2 ? 1 : -1;
+    protected double getSign(int n) {
+        return 0 == n % 2 ? 1. : -1.;
     }
 }
